ASoC:[Backported msm-4.4r9] audio-ext-clk: check for null pointer de-reference
As part of clock test, prepare and unprepare functions are
triggered for all the registered clock's. If clock is
is not used in a flavor of target then pinctrl
pointer is not valid. Hence check for pinctrl before accessing.

CRs-Fixed: 1018208
Change-Id: I3afc0d86476f7a447237180fa411858782d3df1b
Signed-off-by: Yeleswarapu Nagaradhesh <nagaradh@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>
(cherry picked from commit 40aea1da6b129e5acd250766c94964207b11cde0)